updating css for spacing
the height of the sections should not be restricted and should "inherit" their parent's height. The height of the Pic Containers are what need to be restricted. This allows for the sections to be as large as the items within them and the pic to be correctly proportional to the view height